empty files line endings
when a user creates a new file, uses user preferences
however when opening an empty file where notepad++ can't detect the line endings, it uses windows line endings

I've proposed a coupe of changes. First noticing that on construction Buffer::_format is set to the users eol preference, theres no need for a default fallback to be set. 

1) FileManager::loadFileData sets *pFormat to -1 when it can't detect a line ending (empty file)
2) FileManager::loadFile and FileManager::ReloadBuffer only set the line ending of the buffer when format!=-1, thus uses the users preference