Enhance "Close all but This" functionality
Old behaviour:
While "Close all but This" operation, if user answer Yes or No for some files to Save dialog, then cancel for a file to save dialog, all files will be kept.

New behaviour:
The files for which user has answered for saving will be close, even the cancel answer at the end.

Fix #7553